TR 3917

+++ qcmonthly_reports.sh
@@ -16,10 +16,14 @@ umask 002
 
 foreach i ($QCMONTHLY/*.sql)
 
-mv -f $QCREPORTOUTPUTDIR/`basename $i`.[0-9]*.rpt $QCALLELEARCHIVE
-rm -rf $QCREPORTOUTPUTDIR/`basename $i`.current.rpt
-reportisql.csh $i $QCREPORTOUTPUTDIR/`basename $i`.${DATE}.rpt $DSQUERY $MGD
-ln -s $QCREPORTOUTPUTDIR/`basename $i`.${DATE}.rpt $QCREPORTOUTPUTDIR/`basename $i`.current.rpt
+if ( $i == "$QCMONTHLY/ALL_Progress.sql" ) then
+	mv -f $QCREPORTOUTPUTDIR/`basename $i`.[0-9]*.rpt $QCALLELEARCHIVE
+	rm -rf $QCREPORTOUTPUTDIR/`basename $i`.current.rpt
+	reportisql.csh $i $QCREPORTOUTPUTDIR/`basename $i`.${DATE}.rpt $DSQUERY $MGD
+	ln -s $QCREPORTOUTPUTDIR/`basename $i`.${DATE}.rpt $QCREPORTOUTPUTDIR/`basename $i`.current.rpt
+else
+	reportisql.csh $i $QCREPORTOUTPUTDIR/`basename $i`.rpt $DSQUERY $MGD
+endif
 
 end
